Here is the start.
http://www.instructables.com/id/How-to-Make-LEDs-Flash-to-Music-with-an-Arduino/?ALLSTEPS

Here is a website I found to give us a great start for our project. I
followed the build instructions correctly as well got Arduino and
Processing software set up properly with folders after that we just need
to run the cs207 project file and we are off and running.